Separate the oauth definitions and capability
