+++ plus/capabilities/oauth/oauth.go
@@ -3,11 +3,7 @@ package oauth
 import (
 	"errors"
 	"net/http"
-	"plugin"
 	"time"
-
-	"github.com/cloudradar-monitoring/rport/plus/validator"
-	"github.com/cloudradar-monitoring/rport/share/logger"
 )
 
 var (
@@ -123,48 +119,4 @@ type Config struct {
 	RoleClaim     string `mapstructure:"role_claim"`
 	RequiredRole  string `mapstructure:"required_role"`
 	UsernameClaim string `mapstructure:"username_claim"`
-}
-
-const (
-	InitOAuthCapabilityEx  = "InitOAuthCapabilityEx"
-	GitHubOAuthProvider    = "github"
-	MicrosoftOAuthProvider = "microsoft"
-	GoogleOAuthProvider    = "google"
-	Auth0OAuthProvider     = "auth0"
-
-	DefaultLoginURI       = "/oauth/login"
-	DefaultDeviceLoginURI = "/oauth/login/device"
-)
-
-// Capability is used by rportd to maintain loaded info about the plugin's
-// oauth capability
-type Capability struct {
-	Provider CapabilityEx
-
-	Config *Config
-	Logger *logger.Logger
-}
-
-// GetInitFuncName gets the name of the capability init func
-func (cap *Capability) GetInitFuncName() (name string) {
-	return InitOAuthCapabilityEx
-}
-
-// InitProvider invokes the capability init func in the plugin and saves
-// the returned capability provider interface. This interface provides
-// the functions of the capability.
-func (cap *Capability) InitProvider(initFn plugin.Symbol) {
-	fn := initFn.(func(cap *Capability) (capProvider CapabilityEx))
-	cap.Provider = fn(cap)
-}
-
-// GetOAuthCapabilityEx returns the interface to the capability functions
-func (cap *Capability) GetOAuthCapabilityEx() (capEx CapabilityEx) {
-	return cap.Provider
-}
-
-// GetConfigValidator returns a validator interface that can be called to
-// validate the capability config
-func (cap *Capability) GetConfigValidator() (v validator.Validator) {
-	return cap.Provider
 }
